+++ wiring_ex/ant/inc/wiring_ex_lorawan.h
@@ -2,7 +2,7 @@
 #define  WIRING_EX_LORAWAN_H_
 
 #include "lmic.h"
-#include "hal/hal.h"
+#include "hal.h"
 
 #endif /* WIRING_LORAWAN_H_ */